refactor: split non-ts artifact helpers
