+++ .sampo/changesets/issue-540-non-ts-artifact-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split built-in non-TypeScript artifact emitters into focused family and render helper modules while preserving the public scaffold artifact facade.
